Flatbuffer reader

+++ PotreeConverter/src/BINPointReader.cpp
@@ -53,7 +53,7 @@ void BINPointReader::close(){
 	}
 }
 
-long long BINPointReader::numPoints(){
+int64_t BINPointReader::numPoints(){
 	//TODO
 
 	return 0;
